Move dns confirm fragment to fragment package
